skills/visionik/ouracli/pyproject.toml
[project]
name = "ouracli"
version = "0.1.0"
description = "CLI tool for accessing Oura Ring data"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "requests>=2.31.0",
    "typer[all]>=0.9.0",
    "pandas>=2.0.0",
    "python-dotenv>=1.0.0",
    "tabulate>=0.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "hypothesis>=6.92.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
    "types-requests>=2.31.0",
]

[project.scripts]
ouracli = "ouracli.cli:main"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
addopts = "--cov=src --cov-report=html --cov-report=term-missing"

[tool.coverage.run]
omit = ["*/tests/*", "*/venv/*", "*/.venv/*"]

[tool.coverage.report]
fail_under = 75

[tool.black]
line-length = 100

[tool.isort]
profile = "black"
line_length = 100

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "C4", "T10", "PIE", "PT", "RET", "SIM"]
ignore = ["DTZ005", "DTZ007", "B008"]  # Ignore timezone warnings and Typer defaults

[tool.ruff.lint.per-file-ignores]
"src/ouracli/formatters.py" = ["E501"]  # Allow long lines in JS/CSS string literals

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
